added pre-commit hooks
