--- .coveragerc
@@ -1,16 +1,0 @@
-# .coveragerc to control coverage.py
-[run]
-branch = True
-
-source = 
-    katapy/
-omit = 
-    *__init__*
-
-[report]
-include = 
-    katapy/*
-omit = 
-    *__init__*
-
-show_missing = True
