--- .github/workflows/main.yml
@@ -1,51 +1,0 @@
-# This is a basic workflow to help you get started with Actions
-
-name: CI/CD
-
-# Controls when the action will run. Triggers the workflow on push or pull request
-# events but only for the master branch
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
-
-# A workflow run is made up of one or more jobs that can run sequentially or in parallel
-jobs:
-  # This workflow contains a single job called "build"
-  build:
-    # The type of runner that the job will run on
-    runs-on: ubuntu-latest
-
-    # Steps represent a sequence of tasks that will be executed as part of the job
-    steps:
-    # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
-    - uses: actions/checkout@v2
-    - uses: actions/setup-python@v1
-      with:
-        python-version: 3.8
-    - name: Setup Poetry
-      uses: Gr1N/setup-poetry@v2
-    - name: Install dependencies
-      run: poetry install --no-root
-    - name: Run tests and generate coverage
-      run: poetry run coverage run --branch -m pytest && poetry run coverage xml
-    - name: pylint
-      run: poetry run pylint -r n -f parseable --exit-zero katapy > pylint.txt
-    - name: Run Sonar scanner
-      env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
-      run: |
-        export SONAR_SCANNER_VERSION=4.4.0.2170
-        export SONAR_SCANNER_HOME=$HOME/.sonar/sonar-scanner-$SONAR_SCANNER_VERSION-linux
-        curl --create-dirs -sSLo $HOME/.sonar/sonar-scanner.zip https://binaries.sonarsource.com/Distribution/sonar-scanner-cli/sonar-scanner-cli-$SONAR_SCANNER_VERSION-linux.zip
-        unzip -o $HOME/.sonar/sonar-scanner.zip -d $HOME/.sonar/
-        export PATH=$SONAR_SCANNER_HOME/bin:$PATH
-        export SONAR_SCANNER_OPTS="-server"
-        poetry run sonar-scanner \
-        -Dsonar.organization=ethnhll \
-        -Dsonar.projectKey=ethnhll_katapy \
-        -Dsonar.sources=. \
-        -Dsonar.host.url=https://sonarcloud.io
-
