+++ .pre-commit-config.yaml
@@ -1,0 +1,46 @@
+---
+repos:
+    - repo: https://github.com/pre-commit/pre-commit-hooks
+      rev: v4.5.0
+      hooks:
+          - id: check-toml
+          - id: check-yaml
+          - id: end-of-file-fixer
+          - id: trailing-whitespace
+    - repo: https://github.com/python-poetry/poetry
+      rev: 1.7.1
+      hooks:
+          - id: poetry-check
+          - id: poetry-lock
+          - id: poetry-export
+          - id: poetry-install
+    - repo: https://github.com/jumanjihouse/pre-commit-hook-yamlfmt
+      rev: 0.2.3
+      hooks:
+          - id: yamlfmt
+    - repo: https://github.com/adrienverge/yamllint.git
+      rev: v1.33.0
+      hooks:
+          - id: yamllint
+    - repo: https://github.com/psf/black
+      rev: 23.12.0
+      hooks:
+          - id: black
+    - repo: https://github.com/pycqa/flake8
+      rev: 6.1.0
+      hooks:
+          - id: flake8
+    # NOTE: pylint can't be used as a hook against a remote repo
+    # https://pylint.pycqa.org/en/latest/user_guide/installation/pre-commit-integration.html
+    - repo: local
+      hooks:
+          - id: pylint
+            name: pylint
+            entry: pylint
+            language: system
+            types: [python]
+    - repo: https://github.com/PyCQA/bandit
+      rev: 1.7.6
+      hooks:
+          - id: bandit
+            args: [--exclude, tests]
