--- katapy.egg-info/PKG-INFO
@@ -1,11 +1,0 @@
-Metadata-Version: 1.2
-Name: katapy
-Version: 0.1.0
-Summary: UNKNOWN
-Home-page: UNKNOWN
-Author: Ethan Hill
-Author-email: 5614778+ethnhll@users.noreply.github.com
-License: UNKNOWN
-Description: UNKNOWN
-Platform: UNKNOWN
-Requires-Python: >=3.8,<4.0
