+++ pyproject.toml
@@ -5,13 +5,22 @@ description = ""
 authors = ["Ethan Hill <5614778+ethnhll@users.noreply.github.com>"]
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.9"
+typer = "^0.9.0"
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.0.1"
 coverage = "^5.2.1"
-pylint = "^2.5.3"
+
+[tool.poetry.group.dev.dependencies]
+pre-commit = "^3.6.0"
+pylint = "^3.0.3"
 
 [build-system]
 requires = ["poetry>=0.12"]
 build-backend = "poetry.masonry.api"
+
+[tool.pytest.ini_options]
+addopts = [
+    "--import-mode=importlib",
+]
