+++ src/katapy/main.py
@@ -1,0 +1,15 @@
+"""
+The CLI entry-point for this practice tool.
+"""
+
+
+def cli_entry():
+    """
+    Does nothing at the moment...
+    :return:
+    """
+    return None
+
+
+if __name__ == "__main__":
+    cli_entry()
