--- tests/test_katapy.py
@@ -1,5 +1,0 @@
-from katapy import __version__
-
-
-def test_version():
-    assert __version__ == '0.1.0'
